+++ docs/changelog.md
@@ -1,3 +1,12 @@
+## 1.0.4 (2016-10-29)
+
+### Bug Fixes
+* fix validity check for tags (as children) and attributes
+* skip over illegal attributes; add option to fail tag parsing on illegal attribute.
+
+### Features
+* add some tests
+
 ## 1.0.3 (2016-10-22)
 
 ### Bug Fixes